painel-chips-saas/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
bcrypt==4.1.2
pandas